src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile src/requirements.in -o src/requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
base-common==0.3
    # via -r src/requirements.in
base-common-drf==0.4
    # via -r src/requirements.in
brotli==1.1.0
    # via whitenoise
certifi==2024.8.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via rq
concurrent-log-handler==0.9.25
    # via -r src/requirements.in
django==4.2.16
    # via
    #   -r src/requirements.in
    #   base-common
    #   base-common-drf
    #   django-autocomplete-light
    #   django-cas-ng
    #   django-cors-headers
    #   django-filter
    #   django-js-asset
    #   django-jsonform
    #   django-mass-edit
    #   django-redis
    #   django-rq
    #   django-tinymce
    #   django-versatileimagefield
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   sentry-sdk
django-autocomplete-light==3.11.0
    # via -r src/requirements.in
django-cas-ng==5.0.1
    # via -r src/requirements.in
django-cleanup==8.1.0
    # via -r src/requirements.in
django-cors-headers==4.4.0
    # via -r src/requirements.in
django-environ==0.11.2
    # via -r src/requirements.in
django-filter==24.2
    # via -r src/requirements.in
django-js-asset==2.2.0
    # via django-mptt
django-jsonform==2.23.1
    # via -r src/requirements.in
django-mass-edit==3.5.0
    # via -r src/requirements.in
django-mptt==0.16.0
    # via -r src/requirements.in
django-ordered-model==3.7.4
    # via -r src/requirements.in
django-redis==5.4.0
    # via -r src/requirements.in
django-rq==3.0.0
    # via -r src/requirements.in
django-tinymce==4.1.0
    # via -r src/requirements.in
django-versatileimagefield==3.1
    # via -r src/requirements.in
djangorestframework==3.14.0
    # via
    #   -r src/requirements.in
    #   base-common-drf
    #   drf-spectacular
drf-spectacular==0.27.2
    # via
    #   -r src/requirements.in
    #   base-common-drf
drf-spectacular-sidecar==2024.7.1
    # via drf-spectacular
gunicorn==22.0.0
    # via -r src/requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
lxml==5.3.0
    # via
    #   python-cas
    #   python-pptx
packaging==24.1
    # via gunicorn
pillow==11.0.0
    # via
    #   django-versatileimagefield
    #   python-pptx
portalocker==2.10.1
    # via concurrent-log-handler
psutil==6.1.0
    # via rainbow-saddle
psycopg==3.2.1
    # via -r src/requirements.in
psycopg-binary==3.2.1
    # via psycopg
python-cas==1.6.0
    # via django-cas-ng
python-magic==0.4.27
    # via django-versatileimagefield
python-pptx==1.0.2
    # via -r src/requirements.in
pytz==2024.2
    # via djangorestframework
pyyaml==6.0.2
    # via drf-spectacular
rainbow-saddle==0.4.0
    # via -r src/requirements.in
redis==5.1.1
    # via
    #   django-redis
    #   django-rq
    #   rq
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r src/requirements.in
    #   python-cas
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
rq==2.0.0
    # via django-rq
sentry-sdk==2.8.0
    # via -r src/requirements.in
shortuuid==1.0.13
    # via base-common
six==1.16.0
    # via python-cas
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via
    #   psycopg
    #   python-pptx
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via
    #   requests
    #   sentry-sdk
uv==0.2.24
    # via -r src/requirements.in
whitenoise==6.7.0
    # via -r src/requirements.in
xlsxwriter==3.2.0
    # via python-pptx
